Add ShareStoreTask
